test: separate task api contract integration coverage
